// src/version2/models/updatedProjectCategory.ts
/**
 * A project category. */
export interface UpdatedProjectCategory {
  /** The URL of the project category. */
  self?: string;
  /** The ID of the project category. */
  id?: string;
  /** The name of the project category. */
  description?: string;
  /** The description of the project category. */
  name?: string;
}